refactor(speed): separate keyframe timing edits
